refactor: SpamWord의 Dto를 dto 패키지로 이동
SpamWordCreateDto를 dto 패키지로 이동시켰습니다.
